refactor: merge VimMode into the api abstraction
