+++ core/modes/switch_modes.py
@@ -1,6 +1,6 @@
 from talon import Module
 
-from ..rpc.modes import VimMode
+from ..rpc.api import VimAPI
 
 mod = Module()
 
@@ -11,82 +11,82 @@ mod = Module()
 class Actions:
     def vim_set_normal(auto: bool = True):
         """set normal mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_normal_mode(auto=auto)
         return v
 
     def vim_set_normal_exterm():
         """set normal mode, escaping from the terminal mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_normal_mode_exterm()
         return v
 
     def vim_set_normal_np(auto: bool = True):
         """set normal mode and don't preserve the previous mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_normal_mode_np(auto=auto)
         return v
 
     def vim_set_visual():
         """set visual mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_visual_mode()
 
     def vim_set_visual_line():
         """set visual line mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_visual_line_mode()
         return v
 
     def vim_set_visual_block():
         """set visual block mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_visual_block_mode()
 
     def vim_set_insert():
         """set insert mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_insert_mode()
         return v
 
     def vim_set_terminal():
         """set terminal mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_terminal_mode()
         return v
 
     def vim_set_command():
         """set command line mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_command_mode()
         return v
 
     def vim_set_command_exterm():
         """set command line mode, escaping from the terminal mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_command_mode_exterm()
         return v
 
     def vim_set_replace():
         """set replace mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_replace_mode()
         return v
 
     def vim_set_visual_replace():
         """set visual replace mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_visual_replace_mode()
         return v
 
     def vim_set_any_motion_mode():
         """set any motion mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_any_motion_mode()
         return v
 
     def vim_set_any_motion_mode_exterm():
         """set any motion mode, escaping from the terminal mode"""
-        v = VimMode()
+        v = VimAPI()
         v.set_any_motion_mode_exterm()
         return v
