+++ core/rpc/constants.py
@@ -1,0 +1,18 @@
+class VimConstants:
+    """Generic constants"""
+
+    normal_mode_indicators = [
+        "n",
+        "no",
+        "nov",
+        "noV",
+        "no^V",
+        "niI",
+        "niR",
+        "niV",
+    ]
+    visual_mode_indicators = ["v", "V", "^V"]
+    insert_mode_indicators = ["i", "ic", "ix"]
+    terminal_mode_indicators = ["t"]
+    command_mode_indicators = ["c"]
+    replace_mode_indicators = ["R", "Rv", "Rx", "Rc"]
